phase_3/backend/pyproject.toml
# Task: T002, T006 | Spec: @specs/001-sdd-initialization/plan.md
# Description: FastAPI project initialization with ruff configuration
# Purpose: Backend dependencies and Python formatter/linter setup

[project]
name = "phase-2-todo-api"
version = "1.0.0"
description = "Phase 2 Full-Stack Todo Application Backend API"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Phase 2 Development Team", email = "dev@phase2.local"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Task: T006 | Spec: Constitution VI - Code Quality & Formatting
line-length = 100
target-version = "py311"

# Enable strict type checking
extend-select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "alembic",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]
